requirements.txt
# ── 웹 프레임워크 ──────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── DB ────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# ── 환경변수 ──────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Celery / Redis ────────────────────────────────
celery[redis]==5.4.0
redis==5.2.1

# ── AI ───────────────────────────────────────────
google-genai>=1.0.0
anthropic==0.40.0

# ── 인스타그램 다운로드 ──────────────────────────
instaloader==4.13.1

# ── 영상 편집 ────────────────────────────────────
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==11.0.0

# ── HTTP 클라이언트 ───────────────────────────────
httpx==0.28.1
aiofiles==24.1.0

# ── 유틸리티 ─────────────────────────────────────
structlog==24.4.0
tenacity==9.0.0
